Move resultHandler
