+++ core/src/main/java/nl/nn/adapterframework/extensions/javascript/JavascriptEngine.java
@@ -16,7 +16,7 @@
 package nl.nn.adapterframework.extensions.javascript;
 
 import nl.nn.adapterframework.core.PipeLineSession;
-import nl.nn.adapterframework.util.flow.graphviz.ResultHandler;
+import nl.nn.adapterframework.util.flow.ResultHandler;
 import nl.nn.adapterframework.core.ISender;
 
 /**
