+++ core/src/main/java/nl/nn/adapterframework/extensions/javascript/Nashorn.java
@@ -31,7 +31,7 @@ import nl.nn.adapterframework.functional.ThrowingFunction;
 import nl.nn.adapterframework.stream.Message;
 import nl.nn.adapterframework.util.ClassUtils;
 import nl.nn.adapterframework.util.Misc;
-import nl.nn.adapterframework.util.flow.graphviz.ResultHandler;
+import nl.nn.adapterframework.util.flow.ResultHandler;
 
 public class Nashorn implements JavascriptEngine<ScriptEngine> {
