+++ core/src/main/java/nl/nn/adapterframework/util/flow/ResultHandler.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package nl.nn.adapterframework.util.flow.graphviz;
+package nl.nn.adapterframework.util.flow;
 
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
